aaw/aaw_ms/aaw_ms/serializers/category_serializer.py
from rest_framework import serializers
from aaw_ms.models.category_model import Category

class CategorySerializer(serializers.ModelSerializer):

    class Meta:
        model = Category
        fields = ['id', 'name', 'description']
